// src/types.ts
export type User = {
  id: string
  username: string
  name: string
  email: string
  image: string
  token: string
  salt: string
  role: UserRole
  chipsAmount: number
  createdAt: Date
  updatedAt: Date
}
// ...
enum BankActionStatus {
  ХҮЛЭЭЖБАЙНА = 'ХҮЛЭЭЖ БАЙНА',
  АМЖИЛТ = 'АМЖИЛТ',
  БҮТЭЭЛГҮЙ = 'БҮТЭЭЛГҮЙ',
}

export type Recharge = {
  id: string
  amount: number
  bankId: string
  bank?: Bank
  status: BankActionStatus
  createdAt: Date
}

export type Withdraw = {
  id: string
  amount: number
  bankId: string
  bank?: Bank
  status: BankActionStatus
  createdAt: Date
}

export type Bank = {
  id: string
  userId: string
  user?: User
  cardNumber: string
  securityCode: string
  cardHolderName: string
  expiryDate: Date
  username?: string ///// New

  recharges?: Recharge[]
  withdraws?: Withdraw[]

  createdAt: Date
  updatedAt: Date
}